Move agents-dir helpers to hosts.common; sweep manual constructions repo-wide
Relocate get_agents_root_dir and get_agent_state_dir_path from
imbue.mngr.hosts.host to the lower-level imbue.mngr.hosts.common, which is
importable without circular-import risk (its deps -- interfaces.host,
config.data_types, config.agent_config_registry, primitives -- never import back
into hosts.host or agents.base_agent). host.py re-imports and uses both
internally, so the existing imbue.mngr.hosts.host import path still resolves.

This lets base_agent.py (which could not import hosts.host without a cycle) and
every other module route through the shared helpers. Swept the hand-written
host_dir / "agents"[ / str(agent_id)] constructions across mngr, mngr_claude,
mngr_recursive, mngr_ttyd, mngr_claude_usage, mngr_imbue_cloud,
mngr_claude_subagent_proxy, mngr_wait, and minds to use them. Where the id is a
bare str (subagent_wait), used get_agents_root_dir to preserve exact behavior.
No behavior change.
