+++ apps/minds/changelog/mngr-adopt-preserved.md
@@ -1,0 +1,1 @@
+Internal: routed the agents-root directory construction through the shared `get_agents_root_dir` helper (now in `imbue.mngr.hosts.common`). No behavior change.
