+++ libs/mngr_imbue_cloud/changelog/mngr-adopt-preserved.md
@@ -1,0 +1,1 @@
+Internal: routed the agent `data.json` path constructions through the shared `get_agent_state_dir_path` helper (now in `imbue.mngr.hosts.common`). No behavior change.
