+++ libs/mngr_recursive/changelog/mngr-adopt-preserved.md
@@ -1,0 +1,1 @@
+Internal: routed the agent state-dir path construction through the shared `get_agent_state_dir_path` helper (now in `imbue.mngr.hosts.common`). No behavior change.
